fixed bug of cookie invalidation when unauthorized user is logged

+++ src/home/controller/Book.java
@@ -18,7 +18,8 @@ import java.util.HashMap;
 
 public class Book {
 
-    private Book() {}
+    private Book() {
+    }
 
     public static void showBook(HttpServletRequest request, HttpServletResponse response) {
         /**
@@ -27,7 +28,9 @@ public class Book {
 
         DAOFactory sessionDAOFactory = null; //per i cookie
         DAOFactory daoFactory = null; //per il db
+        UserDAO userDAO = null;
         User loggedUser = null;
+        boolean cookieValid = true;
 
 
         try {
@@ -52,8 +55,26 @@ public class Book {
             /* Inizio la transazione sul Database*/
             daoFactory.beginTransaction();
 
+            userDAO = daoFactory.getUserDAO();
 
-            commonView(daoFactory,loggedUser,request);
+            /* controllo lo stato dell'utente */
+            if (loggedUser != null) {
+                /* c'è un utente loggato */
+                if (!sessionUserDAO.isValid(userDAO.findById(loggedUser.getId()))) {
+                    /* utente non autorizzato, invalido il cookie */
+                    System.out.println("UTENTE NON AUTORIZZATO !");
+                    home.controller.Home.logout(request, response);
+                    cookieValid = false;
+                }
+            } else {
+                /*TODO: redirigere a una pagina di errore se si sta provando ad accedere ad una pagina di area riservata senza essere loggati */
+            }
+
+            /* verifico se devo eseguire la logica di business o meno */
+            if (cookieValid) {
+                /* Chiamo la commonView */
+                commonView(daoFactory, loggedUser, request);
+            }
 
             /* Commit della transazione sul db */
             daoFactory.commitTransaction();
@@ -64,25 +85,32 @@ public class Book {
 
         } catch (Exception e) {
             try {
+                if (daoFactory != null) daoFactory.rollbackTransaction(); /* Rollback sul db*/
                 if (sessionDAOFactory != null) sessionDAOFactory.rollbackTransaction();/* Rollback fittizio */
+                System.err.println("ROLLBACK DELLA TRANSAZIONE AVVENUTO CON SUCCESSO");
             } catch (Throwable t) {
+                System.err.println("ERRORE NEL COMMIT/ROLLBACK DELLA TRANSAZIONE");
+
             }
             throw new RuntimeException(e);
 
         } finally {
             try {
+                if (daoFactory != null) daoFactory.closeTransaction(); /* Close sul db*/
                 if (sessionDAOFactory != null) sessionDAOFactory.closeTransaction();/* Close fittizia */
+                System.err.println("CHIUSURA DELLA TRANSAZIONE AVVENUTA CON SUCCESSO");
             } catch (Throwable t) {
             }
         }
 
-
-        /* 1) Attributo che indica se è loggato oppure no */
-        request.setAttribute("loggedOn", loggedUser != null);
-        /* 2) Attributo che indica quale utente è loggato ( da leggere solo se loggedOn = true */
-        request.setAttribute("loggedUser", loggedUser);
-        /* 5) Setto quale view devo mostrare */
-        request.setAttribute("viewUrl", "customer/book");
+        if (cookieValid) {
+            /* 1) Attributo che indica se è loggato oppure no */
+            request.setAttribute("loggedOn", loggedUser != null);
+            /* 2) Attributo che indica quale utente è loggato ( da leggere solo se loggedOn = true */
+            request.setAttribute("loggedUser", loggedUser);
+            /* 5) Setto quale view devo mostrare */
+            request.setAttribute("viewUrl", "customer/book");
+        }
     }
 
     public static void bookAppointment(HttpServletRequest request, HttpServletResponse response) {
@@ -105,6 +133,7 @@ public class Book {
 
         String applicationMessage = "An error occurred!"; /* messaggio da mostrare a livello applicativo ritornato dai DAO */
         boolean inserted = false;
+        boolean cookieValid = true;
 
         try {
             /* Inizializzo il cookie di sessione */
@@ -122,11 +151,6 @@ public class Book {
             /* Controllo se è presente un cookie di sessione tra quelli passati dal browser */
             loggedUser = sessionUserDAO.findLoggedUser();
 
-            /* Fetching dei parametri provenienti dal form di inserimento e salvataggio nelle variabili locali */
-            date = LocalDate.parse(request.getParameter("selected_date"));
-            time = Time.valueOf(request.getParameter("selected_time"));
-            System.err.println("VALORE DI TIME: " + time + " VALORE DI DATE: " + date);
-
             /* DAOFactory per manipolare i dati sul DB */
             daoFactory = DAOFactory.getDAOFactory(Configuration.DAO_IMPL, null);
 
@@ -137,40 +161,63 @@ public class Book {
 
             user = userDAO.findById(loggedUser.getId());
 
-            bookingDAO = daoFactory.getBookingDAO();
 
-            /* Prendo l'unica struttura */
-
-            structureDAO = daoFactory.getStructureDAO();
-
-            structure = structureDAO.fetchStructure();
-
-            System.err.println("VALORE DATE: " + date + " VALORE TIME " + time + " Valore USER " + user.getId() + " VALORE STRUCTURE " + structure.getId());
-
-            /* Effettuo l'inserimento del nuovo appuntamento */
-            try {
-                bookingDAO.insert(date, time, user, structure);
-                inserted = true; /* Se non viene sollevata l'eccezione, l'appuntamento è stato inserito correttamente*/
-            } catch (DuplicatedObjectException e) {
-                applicationMessage = e.getMessage();
-                e.printStackTrace();
+            /* controllo lo stato dell'utente */
+            if (loggedUser != null) {
+                /* c'è un utente loggato */
+                if (!sessionUserDAO.isValid(user)) {
+                    /* utente non autorizzato, invalido il cookie */
+                    System.out.println("UTENTE NON AUTORIZZATO !");
+                    home.controller.Home.logout(request, response);
+                    cookieValid = false;
+                }
+            } else {
+                /*TODO: redirigere a una pagina di errore se si sta provando ad accedere ad una pagina di area riservata senza essere loggati */
             }
 
-            /* Setto l'attributo alreadyBooked */
-            commonView(daoFactory,loggedUser,request);
+            /* verifico se devo eseguire la logica di business o meno */
+            if (cookieValid) {
+                /* Eseguo la logica di business */
+                bookingDAO = daoFactory.getBookingDAO();
 
+                /* Prendo l'unica struttura */
+
+                structureDAO = daoFactory.getStructureDAO();
+
+                structure = structureDAO.fetchStructure();
+
+                /* Fetching dei parametri provenienti dal form di inserimento e salvataggio nelle variabili locali */
+                date = LocalDate.parse(request.getParameter("selected_date"));
+                time = Time.valueOf(request.getParameter("selected_time"));
+                System.err.println("VALORE DI TIME: " + time + " VALORE DI DATE: " + date);
+
+                System.err.println("VALORE DATE: " + date + " VALORE TIME " + time + " Valore USER " + user.getId() + " VALORE STRUCTURE " + structure.getId());
+
+                /* Effettuo l'inserimento del nuovo appuntamento */
+                try {
+                    bookingDAO.insert(date, time, user, structure);
+                    inserted = true; /* Se non viene sollevata l'eccezione, l'appuntamento è stato inserito correttamente*/
+                } catch (DuplicatedObjectException e) {
+                    applicationMessage = e.getMessage();
+                    e.printStackTrace();
+                }
+
+                /* Setto l'attributo alreadyBooked */
+                commonView(daoFactory, loggedUser, request);
+
+                if (inserted) {
+                    /* Solo se viene committata la transazione senza errori siamo sicuri che e' stato inserito correttamente .*/
+                    applicationMessage = "Booking inserted SUCCESSFULLY.";
+                }
+            }
 
             /* Commit della transazione sul db */
             daoFactory.commitTransaction();
 
             /* Commit fittizio */
             sessionDAOFactory.commitTransaction();
-            if (inserted) {
-                /* Solo se viene committata la transazione senza errori siamo sicuri che e' stato inserito correttamente .*/
-                applicationMessage = "Booking inserted SUCCESSFULLY.";
-                System.err.println("COMMIT DELLA TRANSAZIONE AVVENUTO CON SUCCESSO");
-            }
 
+            System.err.println("COMMIT DELLA TRANSAZIONE AVVENUTO CON SUCCESSO");
 
         } catch (Exception e) {
             try {
@@ -196,26 +243,27 @@ public class Book {
             }
         }
 
-
-        /* Setto gli attributi della request che verranno processati dalla book.jsp */
-        /* 1) Attributo che indica se è loggato oppure no */
-        request.setAttribute("loggedOn", loggedUser != null);
-        /* 2) Attributo che indica quale utente è loggato ( da leggere solo se loggedOn = true */
-        request.setAttribute("loggedUser", loggedUser);
-        /* 3) il messaggio da visualizzare nella pagina di inserimento solo se non è null */
-        request.setAttribute("applicationMessage", applicationMessage);
-        /* 4) Attributo che indica in quale struttura si trova loggato l'utente */
-        request.setAttribute("structure", structure);
-        /* 5) l'url della pagina da visualizzare dopo aver effettuato l'inserimento ==> viene visualizzato nuovamente il
-         *     form per consentire un nuovo inserimento */
-        request.setAttribute("viewUrl", "customer/book");
-        /* 6) l'attributo booleano result così da facilitare la scelta dei colori nel frontend JSP ( rosso ==> errore, verde ==> successo per esempio )*/
-        if (inserted) {
-            /* SUCCESS */
-            request.setAttribute("result", "success");
-        } else {
-            /* FAIL */
-            request.setAttribute("result", "fail");
+        if (cookieValid) {
+            /* Setto gli attributi della request che verranno processati dalla book.jsp */
+            /* 1) Attributo che indica se è loggato oppure no */
+            request.setAttribute("loggedOn", loggedUser != null);
+            /* 2) Attributo che indica quale utente è loggato ( da leggere solo se loggedOn = true */
+            request.setAttribute("loggedUser", loggedUser);
+            /* 3) il messaggio da visualizzare nella pagina di inserimento solo se non è null */
+            request.setAttribute("applicationMessage", applicationMessage);
+            /* 4) Attributo che indica in quale struttura si trova loggato l'utente */
+            request.setAttribute("structure", structure);
+            /* 5) l'url della pagina da visualizzare dopo aver effettuato l'inserimento ==> viene visualizzato nuovamente il
+             *     form per consentire un nuovo inserimento */
+            request.setAttribute("viewUrl", "customer/book");
+            /* 6) l'attributo booleano result così da facilitare la scelta dei colori nel frontend JSP ( rosso ==> errore, verde ==> successo per esempio )*/
+            if (inserted) {
+                /* SUCCESS */
+                request.setAttribute("result", "success");
+            } else {
+                /* FAIL */
+                request.setAttribute("result", "fail");
+            }
         }
     }
 
@@ -226,6 +274,7 @@ public class Book {
         Long idToDelete = null; /* id del booking ricevuto da cancellare*/
         DAOFactory sessionDAOFactory = null; //per i cookie
         User loggedUser = null;
+        UserDAO userDAO = null;
         DAOFactory daoFactory = null;
         BookingDAO bookingDAO = null; /* DAO Necessario per poter effettuare la cancellazione del booking */
         Booking bookingToDelete = null;
@@ -233,6 +282,8 @@ public class Book {
         LocalDate currentDate = null;
         String applicationMessage = "An error occurred!"; /* messaggio da mostrare a livello applicativo ritornato dai DAO */
         boolean deleted = false;
+        boolean cookieValid = true;
+
 
         try {
             /* Inizializzo il cookie di sessione */
@@ -262,22 +313,46 @@ public class Book {
             /* Inizio la transazione sul Database*/
             daoFactory.beginTransaction();
 
-            /* è necessario prendersi solo il bookingDAO per poter settare il flag DELETED e la DELETED_REASON */
-            bookingDAO = daoFactory.getBookingDAO();
+            userDAO = daoFactory.getUserDAO();
 
-            /* trovo la prenotazione da flaggare come cancellata (uso findBookingById perche' mi interessano
-               le prenotazioni con stato null dato che quelle con 1 non le vedo, e quelle con 0 vengono gestite in sola lettura) */
-            bookingToDelete = bookingDAO.findBookingById(idToDelete);
+            /* controllo lo stato dell'utente */
+            if (loggedUser != null) {
+                /* c'è un utente loggato */
+                if (!sessionUserDAO.isValid(userDAO.findById(loggedUser.getId()))) {
+                    /* utente non autorizzato, invalido il cookie */
+                    System.out.println("UTENTE NON AUTORIZZATO !");
+                    home.controller.Home.logout(request, response);
+                    cookieValid = false;
+                }
+            } else {
+                /*TODO: redirigere a una pagina di errore se si sta provando ad accedere ad una pagina di area riservata senza essere loggati */
+            }
 
-            /* aggiungo i campi che la rendono cancellata dall'utente */
-            bookingToDelete.setDeletedReason(deletedReason);
+            /* verifico se devo eseguire la logica di business o meno */
+            if (cookieValid) {
+                /* Eseguo la logica di business */
+                /* è necessario prendersi solo il bookingDAO per poter settare il flag DELETED e la DELETED_REASON */
+                bookingDAO = daoFactory.getBookingDAO();
 
-            bookingToDelete.setDeleted(true); /* 1 quando viene modificato dal customer */
+                /* trovo la prenotazione da flaggare come cancellata (uso findBookingById perche' mi interessano
+                 le prenotazioni con stato null dato che quelle con 1 non le vedo, e quelle con 0 vengono gestite in sola lettura) */
+                bookingToDelete = bookingDAO.findBookingById(idToDelete);
 
-            deleted = bookingDAO.deleteBooking(bookingToDelete); /* Se non viene sollevata l'eccezione, l'appuntamento è stato cancellato correttamente*/
+                /* aggiungo i campi che la rendono cancellata dall'utente */
+                bookingToDelete.setDeletedReason(deletedReason);
 
-            /* Setto l'attributo alreadyBooked */
-            commonView(daoFactory,loggedUser,request);
+                bookingToDelete.setDeleted(true); /* 1 quando viene modificato dal customer */
+
+                deleted = bookingDAO.deleteBooking(bookingToDelete); /* Se non viene sollevata l'eccezione, l'appuntamento è stato cancellato correttamente*/
+
+                /* Setto l'attributo alreadyBooked */
+                commonView(daoFactory, loggedUser, request);
+
+                if (deleted) {
+                    applicationMessage = "Booking deleted SUCCESSFULLY. We are sorry that you had to cancel your appointment.";
+                }
+
+            }
 
             /* Commit della transazione sul db */
             daoFactory.commitTransaction();
@@ -285,12 +360,7 @@ public class Book {
             /* Commit fittizio */
             sessionDAOFactory.commitTransaction();
 
-            if (deleted) {
-                /* Se l'impiegato è stato  cancellato committo la transazione */
-                System.err.println("COMMIT DELLA TRANSAZIONE AVVENUTO CON SUCCESSO");
-                /* Solo se viene committata la transazione senza errori siamo sicuri che il dipendente sia stato cancellato correttamente .*/
-                applicationMessage = "Booking deleted SUCCESSFULLY. We are sorry that you had to cancel your appointment.";
-            }
+            System.err.println("COMMIT DELLA TRANSAZIONE AVVENUTO CON SUCCESSO");
 
         } catch (Exception e) {
             try {
@@ -315,24 +385,26 @@ public class Book {
             } catch (Throwable t) {
             }
         }
+        if (cookieValid) {
 
-        /* Setto gli attributi della request che verranno processati dalla show-employees.jsp */
-        /* 1) Attributo che indica se è loggato oppure no */
-        request.setAttribute("loggedOn", loggedUser != null);
-        /* 2) Attributo che indica quale utente è loggato ( da leggere solo se loggedOn = true */
-        request.setAttribute("loggedUser", loggedUser);
-        /* 3) il messaggio da visualizzare nella pagina di elenco solo se non è null */
-        request.setAttribute("applicationMessage", applicationMessage);
-        /* 4) l'url della pagina da visualizzare dopo aver effettuato la cancellazione ==> viene visualizzata nuovamente
-         *     la show-bookings.jsp per consentire altre cancellazioni */
-        request.setAttribute("viewUrl", "customer/book");
-        /* 5) l'attributo booleano result così da facilitare la scelta dei colori nel frontend JSP ( rosso ==> errore, verde ==> successo per esempio )*/
-        if (deleted) {
-            /* SUCCESS */
-            request.setAttribute("result", "success");
-        } else {
-            /* FAIL */
-            request.setAttribute("result", "fail");
+            /* Setto gli attributi della request che verranno processati dalla show-employees.jsp */
+            /* 1) Attributo che indica se è loggato oppure no */
+            request.setAttribute("loggedOn", loggedUser != null);
+            /* 2) Attributo che indica quale utente è loggato ( da leggere solo se loggedOn = true */
+            request.setAttribute("loggedUser", loggedUser);
+            /* 3) il messaggio da visualizzare nella pagina di elenco solo se non è null */
+            request.setAttribute("applicationMessage", applicationMessage);
+            /* 4) l'url della pagina da visualizzare dopo aver effettuato la cancellazione ==> viene visualizzata nuovamente
+             *     la show-bookings.jsp per consentire altre cancellazioni */
+            request.setAttribute("viewUrl", "customer/book");
+            /* 5) l'attributo booleano result così da facilitare la scelta dei colori nel frontend JSP ( rosso ==> errore, verde ==> successo per esempio )*/
+            if (deleted) {
+                /* SUCCESS */
+                request.setAttribute("result", "success");
+            } else {
+                /* FAIL */
+                request.setAttribute("result", "fail");
+            }
         }
     }
